# internal/entity/product.go
package entity

import (
	"time"

	"github.com/cadugr/api-produtos/pkg/entity"
	"github.com/pkg/errors"
)

var (
	ErrIDIsRequired    = errors.New("id is required.")
	ErrInvalidID       = errors.New("id is invalid.")
	ErrNameIsRequired  = errors.New("name is required.")
	ErrPriceIsRequired = errors.New("price is required.")
	ErrInvalidPrice    = errors.New("price is invalid.")
)

type Product struct {
	ID        entity.ID `json:"id"`
	Name      string    `json:"name"`
	Price     float64   `json:"price"`
	CreatedAt time.Time `json:"createdAt"`
}

func NewProduct(name string, price float64) (*Product, error) {
	product := &Product{
		ID:        entity.NewID(),
		Name:      name,
		Price:     price,
		CreatedAt: time.Now(),
	}
	err := product.Validate()
	if err != nil {
		return nil, err
	}
	return product, nil
}

func (p *Product) Validate() error {
	if p.ID.String() == "" {
		return ErrIDIsRequired
	}
	if _, err := entity.ParseID(p.ID.String()); err != nil {
		return ErrInvalidID
	}
	if p.Name == "" {
		return ErrNameIsRequired
	}
	if p.Price == 0 {
		return ErrPriceIsRequired
	}
	if p.Price < 0 {
		return ErrInvalidPrice
	}
	return nil
}